Change long long to long

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	unsigned long long fib_numbers[50] = {1, 2};
+	unsigned long fib_numbers[50] = {1, 2};
 	size_t i;
 
 	printf("1, 2");
@@ -14,7 +14,7 @@ int main(void)
 	for (i = 2; i < 50; i++)
 	{
 		fib_numbers[i] = fib_numbers[i - 1] + fib_numbers[i - 2];
-		printf(", %llu", fib_numbers[i]);
+		printf(", %lu", fib_numbers[i]);
 	}
 
 	printf("\n");
